Rewrite and improve path stream docs
