+++ documentation/docs/develop/02-extensions/04-entries/manifest/path_stream/path_stream_display.mdx
@@ -1,47 +1,48 @@
 import CodeSnippet from "@site/src/components/CodeSnippet";
 
-# PathStreamDisplayEntry
+# PathStreamDisplay
 
-`PathStreamDisplayEntry` lets you render path streams to players. A path stream is a
-sequence of positions calculated by a `PathStreamProducer` and visualised by one or
-more `PathStreamDisplay` implementations.
+`PathStreamDisplayEntry` enables the rendering of custom path streams to players.
+A new display is generated for each viewer whenever a stream is recalculated, which occurs at the interval defined by `refreshDuration`.
 
-A new display is generated for each viewer whenever a stream is recalculated. This
-recalculation happens at the interval defined by `refreshDuration`.
+## When to Use It
 
-## When to use it
-
-Use a `PathStreamDisplayEntry` whenever you want full control over how a path
-stream is visualised. This is useful for quest paths or when highlighting group
-members with custom effects.
+Create a `PathStreamDisplayEntry` when you want to customize the path stream visualization.
 
 ## Basic Structure
 
-The entry constructs a `PathStreamProducer` that computes the route and returns a
-`PathStream`. Every `PathStreamDisplay` created for the viewer consumes that
-stream and renders its positions. In practice you only override the methods that
-create the producer and display.
+A `PathStreamDisplayEntry` must expose these two functions:
+
+- `createProducer()` → `PathStreamProducer`
+- `createDisplay()` → `PathStreamDisplay`
+
+A `PathStreamProducer` is responsible for calculating the path to be displayed and tracking which paths are being displayed.
+Customize this if you want to modify the path stream's positioning.
+
+A `PathStreamDisplay` is responsible for rendering the path to the player, receiving a position every tick and displaying it appropriately.
+
+The entry constructs a `PathStreamProducer` that computes the route and returns a `PathStream`.
+Each `PathStreamDisplay` created for a viewer consumes that stream and renders its positions.
+In practice, you primarily override the methods that create the producer and display.
 
 ```
-Player
-  └─ PathStreamDisplayEntry
-        ├─ createDisplay() ──> PathStreamDisplay
-        └─ createProducer() ─> PathStreamProducer
-                                      └─> PathStream
+PathStreamDisplayEntry
+   ├─ createDisplay() ──> PathStreamDisplay
+   └─ createProducer() ─> PathStreamProducer
+                            └─manages─multiple─> PathStream
+                                                  └─forwards─the─position─> PathStreamDisplay
 ```
-
-<CodeSnippet tag="custom_path_stream_display" json={require("../../../../snippets.json")} />
 
 <CodeSnippet tag="path_stream_display_entry" json={require("../../../../snippets.json")} />
 
-The example above uses `LinePathStreamProducer` which moves from block to block
-towards the end position, displaying each step once. It is the simplest producer
-to start with.
+<CodeSnippet tag="custom_path_stream_display" json={require("../../../../snippets.json")} />
+
+The example above uses `LinePathStreamProducer`, which moves from block to block towards the end position, displaying each step once. It's the simplest producer to begin with.
 
 ## Custom Producers
 
-When you need more advanced behaviour implement your own `PathStreamProducer`
-and `PathStream`. Override `refreshPath()` to supply the custom stream logic.
+Implement your own `PathStreamProducer` when you require greater control over the path stream's positioning.
 
 <CodeSnippet tag="custom_path_stream_producer" json={require("../../../../snippets.json")} />
 
+<CodeSnippet tag="custom_path_stream" json={require("../../../../snippets.json")} />
